backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==4.2.16
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
flake8==7.1.1
gunicorn==23.0.0
idna==3.10
mccabe==0.7.0
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
psycopg2==2.9.10
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.10.1
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
shortuuid==1.0.13
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
